+++ avr/cores/megacommand/MCL/MCLGUI.h
@@ -22,7 +22,7 @@ public:
   void draw_horizontal_arrow(uint8_t x, uint8_t y, uint8_t w);
   bool wait_for_confirm(const char *title, const char *text);
   void draw_infobox(const char *line1, const char *line2,
-                    const int line2_offset = 0);
+                    const int8_t line2_offset = 0);
   void draw_vertical_separator(uint8_t x);
   void draw_vertical_scrollbar(uint8_t x, uint8_t n_items, uint8_t n_window,
                                uint8_t n_current);
